refactor: collapse per-source tables into one catalog
Eight tables keyed by ImageSource lived across five modules — labels and
disc geometry in card-template.ts, the enabled-by-default set in
card-config.ts, the debug rows in debug.ts. Adding a source meant finding
all eight, and nothing linked them, so missing one still type-checked.
One SourceSpec row per source instead, so the compiler names every field
a new source still owes.

Knock-on moves, each needed to keep the graph acyclic:

- ImageSource/GallerySource/IMAGE_SOURCES move to sources.ts, so gallery/
  no longer imports a template module and card-template.ts stops being a hub
- formatDate moves to relative-time.ts, the last edge keeping
  card-template -> gallery-controller -> debug -> card-template alive
- debug.ts splits by role: debug-stats.ts (counters, 5 importers) and
  debug-view.ts (overlay, card.ts only). No module needs both halves.

SvsMoonResolver also moves to the top of its file, matching the sdo and
dscovr resolvers.

Behaviour unchanged: 933 tests, 100% coverage.
